TINKERPOP-2135 Fix for closed idle connections
Connections that were closed while sitting idle in the pool are now
correctly removed from the pool. The pool population method also needed
to be changed as it is now able to populate the pool again if only some
connections were removed. The previous assumption here was that the pool
is either completely populated or completely empty.